Added new plugin that fixes an interface augmented in the re-exported "index.ts" file.
