+++ packages/ckeditor5-dev-docs/lib/docs/index.js
@@ -60,8 +60,7 @@ module.exports = async function build( config ) {
 		entryPoints: files,
 		// logLevel: 'Error',
 		blockTags: [
-			'@eventName',
-			'@memberOf'
+			'@eventName'
 		],
 		inlineTags: [
 			'@link',
@@ -75,9 +74,9 @@ module.exports = async function build( config ) {
 			'typedoc-plugin-rename-defaults',
 			require.resolve( '@ckeditor/typedoc-plugins/lib/module-fixer' ),
 			require.resolve( '@ckeditor/typedoc-plugins/lib/symbol-fixer' ),
+			require.resolve( '@ckeditor/typedoc-plugins/lib/interface-augmentation-fixer' ),
 			require.resolve( '@ckeditor/typedoc-plugins/lib/tag-error' ),
 			require.resolve( '@ckeditor/typedoc-plugins/lib/tag-event' ),
-			require.resolve( '@ckeditor/typedoc-plugins/lib/tag-memberof' ),
 			require.resolve( '@ckeditor/typedoc-plugins/lib/tag-observable' ),
 			require.resolve( '@ckeditor/typedoc-plugins/lib/purge-private-api-docs' ),
